wip: replace packer with lazy
